@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-shopify",
-  "version": "23.1.0",
+  "version": "24.0.0-alpha.1",
   "description": "Shopify’s ESLint rules and configs.",
   "keywords": [
     "eslint",
@@ -11,7 +11,7 @@
   ],
   "author": "Chris Sauve <chris.sauve@shopify.com>",
   "engines": {
-    "node": ">=8.9.0"
+    "node": ">=8.10.0"
   },
   "publishConfig": {
     "access": "public"
@@ -62,7 +62,7 @@
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
     "babel-preset-shopify": "^16.5.0",
-    "eslint": "^4.19.1",
+    "eslint": "^5.4.0",
     "eslint-index": "^1.4.0",
     "eslint-plugin-shopify": "file:./.",
     "isparta": "^4.0.0",
@@ -71,35 +71,35 @@
     "prettier": "^1.12.1",
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
-    "typescript": "^2.8.3"
+    "typescript": "^3.0.1"
   },
   "peerDependencies": {
-    "eslint": "<5 >=4.7.1",
+    "eslint": "<6 >5.0.0",
     "prettier": "^1.12.1"
   },
   "dependencies": {
-    "babel-eslint": "8.2.3",
-    "eslint-config-prettier": "2.9.0",
+    "babel-eslint": "9.0.0",
+    "eslint-config-prettier": "3.0.1",
     "eslint-module-utils": "2.1.1",
-    "eslint-plugin-ava": "4.5.1",
-    "eslint-plugin-babel": "4.1.2",
+    "eslint-plugin-ava": "5.1.0",
+    "eslint-plugin-babel": "5.1.0",
     "eslint-plugin-chai-expect": "1.1.1",
     "eslint-plugin-flowtype": "2.41.0",
-    "eslint-plugin-import": "2.9.0",
-    "eslint-plugin-jest": "21.14.1",
-    "eslint-plugin-jsx-a11y": "6.0.3",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-jest": "21.22.0",
+    "eslint-plugin-jsx-a11y": "6.1.1",
     "eslint-plugin-lodash": "2.6.1",
-    "eslint-plugin-mocha": "4.11.0",
-    "eslint-plugin-node": "6.0.1",
+    "eslint-plugin-mocha": "5.2.0",
+    "eslint-plugin-node": "7.0.1",
     "eslint-plugin-prettier": "2.6.0",
-    "eslint-plugin-promise": "3.7.0",
-    "eslint-plugin-react": "7.7.0",
+    "eslint-plugin-promise": "4.0.0",
+    "eslint-plugin-react": "7.11.1",
     "eslint-plugin-sort-class-members": "1.3.1",
-    "eslint-plugin-typescript": "0.10.0",
+    "eslint-plugin-typescript": "0.12.0",
     "merge": "1.2.0",
     "pascal-case": "^2.0.1",
     "pkg-dir": "2.0.0",
     "pluralize": "^7.0.0",
-    "typescript-eslint-parser": "17.0.1"
+    "typescript-eslint-parser": "18.0.0"
   }
 }
