@@ -9,11 +9,8 @@ module.exports = {
   },
 
   parserOptions: {
-    ecmaVersion: 2017,
+    ecmaVersion: 2018,
     sourceType: 'module',
-    ecmaFeatures: {
-      experimentalObjectRestSpread: true,
-    },
   },
 
   plugins: ['babel', 'promise', 'sort-class-members', 'import'],
