@@ -25,6 +25,8 @@ module.exports = {
   eqeqeq: ['error', 'allow-null'],
   // Make sure for-in loops have an if statement
   'guard-for-in': 'error',
+  // enforce a maximum number of classes per file
+  'max-classes-per-file': 'off',
   // Disallow the use of alert, confirm, and prompt
   'no-alert': 'error',
   // Disallow lexical declarations in case clauses
@@ -105,7 +107,7 @@ module.exports = {
   // Disallow use of javascript: urls.,
   'no-script-url': 'off',
   // Disallow assignments where both sides are exactly the same
-  'no-self-assign': 'error',
+  'no-self-assign': ['error', {props: true}],
   // Disallow comparisons where both sides are exactly the same
   'no-self-compare': 'error',
   // Disallow use of comma operator
@@ -138,6 +140,8 @@ module.exports = {
   radix: 'error',
   // Disallow async functions which have no await expression
   'require-await': 'error',
+  // Enforce the use of u flag on RegExp
+  'require-unicode-regexp': 'off',
   // Requires to declare all vars on top of their containing scope
   'vars-on-top': 'off',
   // Require immediate function invocation to be wrapped in parentheses
