@@ -23,8 +23,14 @@ module.exports = {
   'import/no-absolute-path': 'error',
   // Forbid Webpack loader syntax in imports
   'import/no-webpack-loader-syntax': 'error',
+  // Forbid a module from importing itself
+  'import/no-self-import': 'error',
+  // Forbid a module from importing a module with a dependency path back to itself
+  'import/no-cycle': 'error',
   // Ensures that there are no useless path segments
   'import/no-useless-path-segments': 'error',
+  // Forbid importing modules from parent directories
+  'import/no-relative-parent-imports': 'off',
 
   // Helpful warnings
 
@@ -42,8 +48,6 @@ module.exports = {
   'import/no-extraneous-dependencies': 'error',
   // Forbid the use of mutable exports with var or let.
   'import/no-mutable-exports': 'error',
-  // Forbid a module from importing itself
-  'import/no-self-import': 'error',
 
   // Module systems
 
