@@ -1,16 +1,22 @@
 // see https://github.com/sindresorhus/eslint-plugin-ava
 
 module.exports = {
+  // Enforce assertion to be made in a test body
+  'jest/expect-expect': 'error',
   // Disallow all forms of skipped tests (`.skip`, prepending `x` to test helpers, empty function bodies, `pending()` calls.)
   'jest/no-disabled-tests': 'error',
   // Disallow disabling tests via appending `.only` or prepending `f` to test helpers.
   'jest/no-focused-tests': 'error',
   // Make it easier to find failing tests by preventing duplicate test names in a suite.
   'jest/no-identical-title': 'error',
+  // Disallow Jasmine globals
+  'jest/no-jasmine-globals': 'error',
   // Limited snapshot sizes to keep snapshops manageable and reviewable.
   'jest/no-large-snapshots': ['error', {maxSize: 12}],
   // For better failure messages, use `toHaveLength()` to on object lengths.
   'jest/prefer-to-have-length': 'error',
+  // Suggest using toMatchInlineSnapshot()
+  'jest/prefer-inline-snapshots': 'off',
   // For better failure messages, use `toBeNull()` to assert on null values.
   'jest/prefer-to-be-null': 'error',
   // For better failure messages, use `toBeUndefined()` to assert on undefined values.
@@ -27,8 +33,10 @@ module.exports = {
   'jest/consistent-test-it': 'error',
   // Use .only and .skip over f and x
   'jest/no-test-prefixes': 'error',
+  // Disallow explicitly returning from tests
+  'jest/no-test-return-statement': 'error',
   // Enforce lowercase test names
-  'jest/lowercase-name': 'off',
+  'jest/lowercase-name': ['error', {ignore: ['describe']}],
   // Disallow importing Jest
   'jest/no-jest-import': 'error',
   // Disallow vague words in test statements.
