@@ -30,6 +30,16 @@ module.exports = {
   'node/no-extraneous-require': 'error',
   // Enforce either module.exports or exports.
   'node/exports-style': ['error', 'module.exports'],
+  // enforce either Buffer or require("buffer").Buffer
+  'node/prefer-global/buffer': 'error',
+  //	enforce either console or require("console")
+  'node/prefer-global/console': 'error',
+  // enforce either process or require("process")
+  'node/prefer-global/process': 'error',
+  // enforce either URLSearchParams or require("url").URLSearchParams
+  'node/prefer-global/url-search-params': 'error',
+  // enforce either URL or require("url").URL
+  'node/prefer-global/url': 'error',
   // Disallow deprecated API.
   'node/no-deprecated-api': 'error',
   // Disallow import and export declarations for files that don't exist.
@@ -42,8 +52,12 @@ module.exports = {
   'node/no-unpublished-bin': 'error',
   // Disallow require()s for files that are not published.
   'node/no-unpublished-require': 'off',
-  // Disallow unsupported ECMAScript features on the specified version.
-  'node/no-unsupported-features': 'off',
+  // disallow unsupported ECMAScript built-ins on the specified version
+  'node/no-unsupported-features/es-builtins': 'off',
+  // disallow unsupported ECMAScript syntax on the specified version
+  'node/no-unsupported-features/es-syntax': 'off',
+  // disallow unsupported Node.js built-in APIs on the specified version
+  'node/no-unsupported-features/node-builtins': 'error',
   // If you turn this rule on, ESLint comes to address process.exit() as throw in code path analysis.
   'node/process-exit-as-throw': 'off',
   // Suggest correct usage of shebang.
