@@ -34,7 +34,7 @@ module.exports = {
       return left.type === 'Identifier' ||
       (
         left.type === 'ObjectPattern' && left.properties.some((prop) => {
-          return prop.type === 'SpreadProperty' || prop.type === 'ExperimentalRestProperty';
+          return prop.type === 'SpreadProperty' || prop.type === 'ExperimentalRestProperty' || prop.type === 'RestElement';
         })
       ) ||
       (
